toxicity-detection-poc/services/game-message-proxy/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
kafka-python==2.0.2
pydantic==1.10.7
python-dotenv==1.0.0
